requirements.txt
aiohttp==3.8.3
discord.py==2.0.1
python-dotenv==0.21.0
beautifulsoup4==4.11.1
lxml==4.9.1